Feat:Implement JWT authentication

+++ controller/LikeController.js
@@ -1,42 +1,70 @@
 const { body } = require('express-validator');
 const conn = require('../mariadb');
 const {StatusCodes} = require('http-status-codes');
+const jwt = require('jsonwebtoken');
+const ensureAuthorization = require('../auth');
 
 const addlike = (req, res) => {
     // 카테고리 전체 목록 리스트
     const {liked_book_id} = req.params;
-    const {user_id} = req.body;
 
-    let sql = `INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?);`;
-    let values = [user_id, liked_book_id];
-    conn.query(sql, values,
-        (err, results) => {
-            if (err) {
-                console.log(err);
-                return res.status(StatusCodes.BAD_REQUEST).end()
+    let authorization = ensureAuthorization(req, res);
+
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            message: "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    } 
+    else if(authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            message: "잘못된 토큰입니다."
+        });
+    }
+    else {
+        let sql = `INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?);`;
+        let values = [authorization.id, liked_book_id];
+        conn.query(sql, values,
+            (err, results) => {
+                if (err) {
+                    console.log(err);
+                    return res.status(StatusCodes.BAD_REQUEST).end()
+                }
+                return res.status(StatusCodes.OK).json(results);
             }
-            return res.status(StatusCodes.OK).json(results);
-        }
-    )
+        )
+    }
 }
 
 const removeLike = (req, res) => {
     // 좋아요 제거(취소)
     const {liked_book_id} = req.params;
-    const {user_id} = req.body;
 
-    let sql = `DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?;`;
-    let values = [user_id, liked_book_id];
+    let authorization = ensureAuthorization(req, res);
 
-    conn.query(sql, values,
-        (err, results) => {
-            if (err) {
-                console.log(err);
-                return res.status(StatusCodes.BAD_REQUEST).end();
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            message: "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    } 
+    else if(authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            message: "잘못된 토큰입니다."
+        });
+    }
+    else {
+        let sql = `DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?;`;
+        let values = [authorization.id, liked_book_id];
+
+        conn.query(sql, values,
+            (err, results) => {
+                if (err) {
+                    console.log(err);
+                    return res.status(StatusCodes.BAD_REQUEST).end();
+                }
+                return res.status(StatusCodes.OK).json(results);
             }
-            return res.status(StatusCodes.OK).json(results);
-        }
-    )
+        )
+    }
 }
 
 module.exports = {
